Restore DataBase BackUp
Restoring Database Backup. In this problem it was not possible to perform a TRUNCATE and insert the data again because the system continued entering data. You would have to parse the data that you deleted and re-insert only the deleted data.
"Change the wheel of the car with the car moving." LOL